+++ src/jacobipar.h
@@ -1,6 +1,6 @@
 #ifndef __jacobipar_H__ 
 #define __jacobipar_H__
 
-double* jacobipar(double* matrix,double* constants,int N,float tolerance,int maxIterations);
+float* jacobipar(float* matrix,float* constants,int N,float tolerance,int maxIterations);
 
 #endif
