+++ src/jacobiseq.h
@@ -1,6 +1,6 @@
 #ifndef __jacobiseq_H__ 
 #define __jacobiseq_H__
 
-double* jacobiseq(double* matrix,double* constants,int N,float tolerance,int maxIterations);
+float* jacobiseq(float* matrix,float* constants,int N,float tolerance,int maxIterations);
 
 #endif
